requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
azure-cognitiveservices-speech==1.40.0
ffmpeg-python==0.2.0
pydub==0.25.1
